feat : THe fontend now interactive with backend
